Removing the max-width for now per Erik
Since we're about to move a bunch of the VectorBeta changes to
Vector, it would be good to make sure that people don't think we're
going to apply the max-width (since that was contentious). Erik
suggested that we go ahead and remove it from VectorBeta as well,
although we can add it back later if we decide we want to
experiement with it more.

Change-Id: Ic1c37b736c3315ab68b8dff77e6a74ece79b5d68